microservicio2/requirements.txt
Flask
Flask-Cors
Flask-RESTful
requests
urllib3
marshmallow
flask-swagger-ui
Flask-JWT-Extended
Werkzeug
psycopg2
setuptools
pyjwt
pytz